requirements.txt
langgraph>=0.0.31
langchain-core>=0.1.40
langchain-openai>=0.1.1
openai>=1.0.0
python-dotenv>=1.0.1
ffmpeg-python>=0.2.0
soundfile>=0.12.1
mutagen>=1.47.0
aiohttp>=3.9.3
pydantic>=2.6.4
playwright>=1.48.0
PySide6>=6.7.0
numpy>=1.26.0
crewai>=0.102.0
pyscenedetect[opencv]>=0.6.3
scikit-learn>=1.4.0
